Refactor pdf helper methods into PdfHelper module in spec/support
